Shifted app initialization to a separate file
h/t @jmcarp!
