+++ openfecwebapp/env.py
@@ -1,0 +1,6 @@
+import cfenv
+
+env = cfenv.AppEnv()
+
+__all__ = ['env']
+
